Add rake task to publish models for seeding (#8)
When a new application is setup it needs all previously created
records to have messages sent again for them. This adds a rake
task that will allow you to publish all records for any model
setup with the Publish mixin.

It also includes a touch option to update the updated_at timestamp
on the records so that the receiving systems will not ignore the
messages if they've already received them. This is useful in
some edge cases like if a new column is added in a receiving
system and it needs that column populated.

needed-by #7